Move util methods into new `AuthenticateUpsellViews` class.
